@@ -15,39 +15,46 @@
   "bugs": {
     "url": "https://github.com/infra-geo-ouverte/igo2-lib/issues"
   },
+  "engines": {
+    "node": ">=18.10.0"
+  },
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --proxy-config proxy.conf.json",
+    "start": "ng serve --hmr --proxy-config proxy.conf.json",
     "start.watch": "run-p watch.libs \"_start.watch -- -- {@}\" --",
     "build": "ng build",
     "lint": "ng lint",
     "lint.fix": "ng lint --fix",
-    "e2e": "ng e2e",
-    "postinstall": "gulp fixLibs && ngcc --tsconfig ./tsconfig.json",
+    "e2e": "ng run demo:e2e:production",
+    "e2e.local": "ng run demo:e2e:local",
     "build.all": "npm run build.libs && npm run build.demo",
     "build.libs": "rimraf ./dist && npm run build.utils && npm run build.core && npm run build.common && npm run build.auth && npm run build.geo && npm run build.context && npm run build.integration",
-    "build.utils": "ng build utils --configuration production && gulp core:bundleLocale",
-    "build.core": "ng build core --configuration production && gulp core && gulp core:bundleLocale",
-    "build.common": "ng build common --configuration production && gulp common && gulp core:bundleLocale",
-    "build.auth": "ng build auth --configuration production && gulp auth && gulp core:bundleLocale",
-    "build.geo": "ng build geo --configuration production && gulp geo && gulp core:bundleLocale",
-    "build.context": "ng build context --configuration production && gulp context && gulp core:bundleLocale",
-    "build.integration": "ng build integration --configuration production && gulp integration && gulp core:bundleLocale",
-    "build.demo": "ng build demo --configuration=production && gulp core:bundleLocale",
+    "build.utils": "ng build utils --configuration production",
+    "build.core": "ng build core --configuration production && npm run build.locale",
+    "build.common": "ng build common --configuration production",
+    "build.auth": "ng build auth --configuration production",
+    "build.geo": "ng build geo --configuration production",
+    "build.context": "ng build context --configuration production",
+    "build.integration": "ng build integration --configuration production",
+    "build.demo": "ng build demo --configuration=production",
     "build.ghpages": "ng build demo --configuration=production --output-path ./dist/ghpages --base-href /igo2-lib/",
     "build.doc": "compodoc -p tsconfig.json --disablePrivate --disableProtected --disableInternal --disableLifeCycleHooks --theme material",
+    "build.locale": "ts-node scripts/src/locale-merge.ts",
+    "postbuild.core": "ts-node scripts/src/postbuild-core.ts",
+    "postbuild.common": "ts-node scripts/src/postbuild-common.ts",
+    "postbuild.geo": "ts-node scripts/src/postbuild-geo.ts",
     "copy": "copy-newer --cwd dist ** \"..\\igo2\\node_modules\\@igo2\"",
     "doc": "npm run build.doc && compodoc -s --port 4220",
-    "pretest": "ng lint",
     "test": "ng test --watch=false --browsers=ChromeHeadless",
     "test.watch": "ng test",
-    "test.all": "npm run test.libs && npm run test.demo",
-    "test.libs": "npm run lint && npm run test.core && npm run test.common && npm run test.geo && npm run test.context",
-    "test.core": "ng test core --watch=false --code-coverage",
-    "test.common": "ng test common --watch=false --code-coverage",
-    "test.geo": "ng test geo --watch=false --code-coverage",
-    "test.context": "ng test context --watch=false --code-coverage",
-    "test.demo": "ng test demo --watch=false --code-coverage && ng e2e --port 4300",
+    "test.libs": "npm run test.utils && npm run test.core && npm run test.common && npm run test.auth && npm run test.geo && npm run test.context",
+    "test.auth": "ng test auth --watch=false",
+    "test.common": "ng test common --watch=false",
+    "test.context": "ng test context --watch=false",
+    "test.core": "ng test core --watch=false",
+    "test.geo": "ng test geo --watch=false",
+    "test.utils": "ng test utils --watch=false",
+    "test.demo": "ng test demo --watch=false",
     "watch.libs": "rimraf ./dist && run-p _watch.*",
     "_watch.utils": "ng build utils --watch",
     "_watch.core": "wait-on dist/utils/package.json && ng build core --watch",
@@ -70,122 +76,106 @@
     "version.linux": "gulp bumpVersion && rimraf ./dist && npm run build.all && npm run changelog && git add -u",
     "postversion": "git push && git push --tags && npm run publish && npm run ghpages.win",
     "postversion.linux": "git push && git push --tags && npm run publish && npm run ghpages",
-    "preghpages": "gulp core:bundleLocale && npm run build.ghpages && echo \"include: ['_default.json', '_contexts.json', '_base.json']\" > dist/ghpages/_config.yml",
+    "preghpages": "gulp bundleLocale && npm run build.ghpages && echo \"include: ['_default.json', '_contexts.json', '_base.json']\" > dist/ghpages/_config.yml",
     "ghpages": "VERSION=$(node -p -e \"require('./package.json').version\") && npx ngh --dir=dist/ghpages --no-silent=false --message=$VERSION",
     "postghpages": "rimraf ./dist/ghpages",
-    "preghpages.win": "gulp core:bundleLocale && npm run build.ghpages && echo include: ['_default.json', '_contexts.json', '_base.json'] > dist/ghpages/_config.yml",
+    "preghpages.win": "gulp bundleLocale && npm run build.ghpages && echo include: ['_default.json', '_contexts.json', '_base.json'] > dist/ghpages/_config.yml",
     "ghpages.win": "npx ngh --dir=dist/ghpages --no-silent=false --message=%npm_package_version%",
     "postghpages.win": "rimraf ./dist/ghpages"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^14.3.0",
-    "@angular/cdk": "^14.2.7",
-    "@angular/common": "^14.3.0",
-    "@angular/compiler": "^14.3.0",
-    "@angular/core": "^14.3.0",
-    "@angular/forms": "^14.3.0",
-    "@angular/material": "^14.2.7",
-    "@angular/platform-browser": "^14.3.0",
-    "@angular/router": "^14.3.0",
-    "@azure/msal-angular": "^2.4.1",
-    "@azure/msal-browser": "^2.28.1",
-    "@mat-datetimepicker/core": "~10.0.0",
-    "@mdi/angular-material": "^6.2.95",
-    "@ngx-translate/core": "^14.0.0",
+    "@angular/animations": "^16.1.7",
+    "@angular/cdk": "^16.1.6",
+    "@angular/common": "^16.1.7",
+    "@angular/compiler": "^16.1.7",
+    "@angular/core": "^16.1.7",
+    "@angular/forms": "^16.1.7",
+    "@angular/material": "^16.1.6",
+    "@angular/platform-browser": "^16.1.7",
+    "@angular/platform-browser-dynamic": "^16.1.7",
+    "@angular/router": "^16.1.7",
+    "@azure/msal-angular": "^3.0.0-beta.1",
+    "@mat-datetimepicker/core": "~12.0.0",
+    "@mdi/angular-material": "^7.2.96",
+    "@ngx-translate/core": "^15.0.0",
+    "@ngx-translate/http-loader": "^8.0.0",
     "@turf/buffer": "^6.5.0",
     "@turf/helpers": "^6.5.0",
     "@turf/line-intersect": "^6.5.0",
     "@turf/point-on-feature": "^6.5.0",
-    "ajv": "^6.9.1",
-    "angular-shepherd": "14.0.0",
+    "angular-shepherd": "16.0.0",
     "bowser": "^2.10.0",
     "core-js": "^3.30.2",
     "file-saver": "^2.0.2",
     "flexsearch": "0.7.21",
-    "hammerjs": "^2.0.8",
+    "hammerjs": "2.0.8",
     "html2canvas": "^1.4.1",
     "jspdf": "^2.5.1",
     "jspdf-autotable": "^3.5.29",
     "jszip": "^3.10.1",
     "jwt-decode": "^2.2.0",
     "moment": "^2.29.4",
-    "ngx-color-picker": "^13.0.0",
+    "ngx-color-picker": "^14.0.0",
     "ngx-indexed-db": "^11.0.2",
-    "ngx-toastr": "^15.0.0",
+    "ngx-toastr": "^17.0.0",
     "nosleep.js": "^0.12.0",
-    "ol": "^7.3.0",
-    "proj4": "^2.8.0",
-    "rxjs": "^7.5.6",
-    "scroll-into-view-if-needed": "^2.2.29",
-    "scss-bundle": "^3.1.2",
-    "stream": "^0.0.2",
+    "ol": "^7.4.0",
+    "proj4": "^2.9.0",
+    "rxjs": "^7.8.0",
+    "scroll-into-view-if-needed": "^3.0.0",
     "striptags": "^3.2.0",
     "ts-cacheable": "^1.0.6",
-    "ts-md5": "^1.2.4",
-    "tslib": "^2.4.0",
+    "ts-md5": "^1.3.0",
+    "tslib": "^2.6.0",
     "typy": "^3.3.0",
-    "unorm": "^1.4.1",
     "windows-1252": "^3.0.4",
-    "zone.js": "~0.11.7"
+    "zone.js": "~0.13.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^14.2.11",
-    "@angular-eslint/builder": "14.0.2",
-    "@angular-eslint/eslint-plugin": "14.0.2",
-    "@angular-eslint/eslint-plugin-template": "14.0.2",
-    "@angular-eslint/schematics": "14.0.2",
-    "@angular-eslint/template-parser": "14.0.2",
-    "@angular/platform-browser-dynamic": "^14.3.0",
-    "@angular/cli": "^14.2.10",
-    "@angular/compiler-cli": "^14.3.0",
-    "@babel/core": "^7.18.10",
-    "@babel/preset-env": "^7.18.10",
+    "@angular-devkit/build-angular": "^16.1.6",
+    "@angular-eslint/builder": "^16.1.0",
+    "@angular-eslint/eslint-plugin": "^16.1.0",
+    "@angular-eslint/eslint-plugin-template": "^16.1.0",
+    "@angular-eslint/schematics": "^16.1.0",
+    "@angular-eslint/template-parser": "^16.1.0",
+    "@angular/cli": "^16.1.6",
+    "@angular/compiler-cli": "^16.1.7",
     "@compodoc/compodoc": "^1.1.15",
-    "@ngx-translate/http-loader": "^7.0.0",
-    "@octokit/core": "^4.0.0",
+    "@cypress/schematic": "^2.5.0",
+    "@types/flexsearch": "^0.7.3",
     "@types/geojson": "^7946.0.10",
     "@types/hammerjs": "^2.0.36",
-    "@types/html2canvas": "0.0.35",
-    "@types/jasmine": "^3.5.11",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^14.18.23",
-    "@typescript-eslint/eslint-plugin": "^5.59.6",
-    "@typescript-eslint/parser": "^5.33.0",
-    "@types/flexsearch": "^0.7.3",
+    "@types/html2canvas": "1.0.0",
+    "@types/jasmine": "~4.3.0",
+    "@types/lodash": "^4.14.195",
+    "@types/node": "^18.0.0",
+    "@typescript-eslint/eslint-plugin": "^6.2.0",
+    "@typescript-eslint/parser": "^6.2.0",
     "angular-cli-ghpages": "^1.0.3",
-    "conventional-changelog-cli": "^2.0.12",
+    "conventional-changelog-cli": "^3.0.0",
     "copy-newer": "^2.1.2",
-    "del": "^6.0.0",
-    "eslint": "^8.21.0",
-    "eslint-plugin-unused-imports": "^2.0.0",
+    "cypress": "^12.17.0",
+    "del": "^7.0.0",
+    "eslint": "^8.46.0",
+    "eslint-plugin-cypress": "^2.13.3",
+    "eslint-plugin-unused-imports": "^3.0.0",
     "gulp": "^4.0.0",
-    "gulp-babel": "^8.0.0",
-    "gulp-concat": "^2.6.1",
-    "gulp-exec": "^5.0.0",
-    "gulp-json-editor": "^2.4.2",
-    "gulp-merge-json": "^2.1.1",
-    "gulp-replace": "^1.0.0",
-    "gulp-sequence": "^1.0.0",
-    "gulp-wait": "0.0.2",
-    "jasmine-core": "~3.8.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "^6.4.2",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "^2.2.0",
-    "karma-jasmine": "~4.0.1",
-    "karma-jasmine-html-reporter": "^1.7.0",
-    "ng-packagr": "^14.2.2",
+    "gulp-json-editor": "^2.5.0",
+    "gulp-replace": "^1.1.0",
+    "jasmine-core": "~5.1.0",
+    "jasmine-spec-reporter": "~7.0.0",
+    "karma": "~6.4.2",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "lodash": "^4.17.21",
+    "ng-packagr": "^16.1.0",
     "npm-run-all": "^4.1.5",
-    "protractor": "~7.0.0",
-    "sass": "^1.54.4",
+    "sass": "^1.64.0",
     "ts-node": "^10.9.1",
-    "typescript": "~4.6.4",
-    "wait-on": "^6.0.0",
-    "webdriver-manager": "^12.1.6",
-    "webpack": "^5.76.0"
-  },
-  "engines": {
-    "node": ">=16.19.0"
+    "typescript": "~5.1.0",
+    "wait-on": "^7.0.0"
   }
 }
