@@ -15,24 +15,23 @@
   },
   "dependencies": {
     "jwt-decode": "^2.2.0",
-    "ts-cacheable": "^1.0.5",
-    "ts-md5": "^1.2.4",
-    "tslib": "^2.4.0"
+    "ts-cacheable": "^1.0.6",
+    "ts-md5": "^1.3.0",
+    "tslib": "^2.6.0"
   },
   "peerDependencies": {
-    "@angular/cdk": "^14.2.7",
-    "@angular/common": "^14.3.0",
-    "@angular/core": "^14.3.0",
-    "@angular/forms": "^14.3.0",
-    "@angular/material": "^14.2.7",
-    "@angular/router": "^14.3.0",
+    "@angular/cdk": "^16.1.6",
+    "@angular/common": "^16.1.7",
+    "@angular/core": "^16.1.7",
+    "@angular/forms": "^16.1.7",
+    "@angular/material": "^16.1.6",
+    "@angular/router": "^16.1.7",
     "@igo2/core": "^1.15.4",
     "@igo2/utils": "^1.15.4",
     "rxjs": "^7.5.6",
-    "@azure/msal-angular": "^2.4.1",
-    "@azure/msal-browser": "^2.28.1"
+    "@azure/msal-angular": "^3.0.0-beta.1"
   },
   "engines": {
-    "node": ">=16.19.0"
+    "node": ">=18.10.0"
   }
 }
