@@ -14,24 +14,24 @@
     "url": "https://github.com/infra-geo-ouverte/igo2-lib/issues"
   },
   "peerDependencies": {
-    "@angular/animations": "^14.3.0",
-    "@angular/common": "^14.3.0",
-    "@angular/core": "^14.3.0",
-    "@angular/forms": "^14.3.0",
-    "@angular/material": "^14.2.7",
-    "@angular/platform-browser": "^14.3.0",
+    "@angular/animations": "^16.1.7",
+    "@angular/common": "^16.1.7",
+    "@angular/core": "^16.1.7",
+    "@angular/forms": "^16.1.7",
+    "@angular/material": "^16.1.6",
+    "@angular/platform-browser": "^16.1.7",
     "@igo2/auth": "^1.15.4",
     "@igo2/common": "^1.15.4",
     "@igo2/core": "^1.15.4",
     "@igo2/utils": "^1.15.4",
     "@igo2/geo": "^1.15.4",
-    "ol": "^7.3.0",
-    "rxjs": "^7.5.6"
+    "ol": "^7.4.0",
+    "rxjs": "^7.8.0"
   },
   "dependencies": {
-    "tslib": "^2.4.0"
+    "tslib": "^2.6.0"
   },
   "engines": {
-    "node": ">=16.19.0"
+    "node": ">=18.10.0"
   }
 }
