@@ -13,32 +13,41 @@
   "bugs": {
     "url": "https://github.com/infra-geo-ouverte/igo2-lib/issues"
   },
+  "exports": {
+    "./style": {
+      "sass": "./style/style.scss"
+    },
+    "./theming": {
+      "sass": "./geo-theme.scss"
+    }
+  },
   "dependencies": {
-    "ts-cacheable": "^1.0.5",
-    "ts-md5": "^1.2.4",
+    "ts-cacheable": "^1.0.6",
+    "ts-md5": "^1.3.0",
     "@turf/helpers": "^6.5.0",
     "@turf/line-intersect": "^6.5.0",
     "@turf/point-on-feature": "^6.5.0",
-    "tslib": "^2.4.0",
+    "tslib": "^2.6.0",
     "nosleep.js": "^0.12.0",
     "windows-1252": "^3.0.4"
   },
   "peerDependencies": {
-    "@angular/animations": "^14.3.0",
-    "@angular/common": "^14.3.0",
-    "@angular/core": "^14.3.0",
-    "@angular/forms": "^14.3.0",
-    "@angular/material": "^14.2.7",
-    "@angular/platform-browser": "^14.3.0",
+    "@angular/animations": "^16.1.7",
+    "@angular/common": "^16.1.7",
+    "@angular/core": "^16.1.7",
+    "@angular/forms": "^16.1.7",
+    "@angular/material": "^16.1.6",
+    "@angular/platform-browser": "^16.1.7",
     "@igo2/common": "^1.15.4",
     "@igo2/core": "^1.15.4",
     "@igo2/utils": "^1.15.4",
+    "@mat-datetimepicker/core": "~12.0.0",
     "flexsearch": "0.7.21",
-    "ol": "^7.3.0",
-    "proj4": "^2.8.0",
-    "rxjs": "^7.5.6",
+    "ol": "^7.4.0",
+    "proj4": "^2.9.0",
+    "rxjs": "^7.8.0",
     "striptags": "^3.2.0",
-    "ngx-color-picker": "^13.0.0",
+    "ngx-color-picker": "^14.0.0",
     "ngx-indexed-db": "^11.0.2"
   },
   "optionalDependencies": {
@@ -46,10 +55,9 @@
     "html2canvas": "^1.4.1",
     "jspdf": "^2.5.1",
     "jszip": "^3.10.1",
-    "moment": "^2.29.4",
-    "stream": "^0.0.2"
+    "moment": "^2.29.4"
   },
   "engines": {
-    "node": ">=16.19.0"
+    "node": ">=18.10.0"
   }
 }
