@@ -14,20 +14,18 @@
     "url": "https://github.com/infra-geo-ouverte/igo2-lib/issues"
   },
   "peerDependencies": {
-    "@angular/common": "^14.3.0",
-    "@angular/core": "^14.3.0",
+    "@angular/common": "^16.1.7",
+    "@angular/core": "^16.1.7",
     "@igo2/geo": "1.15.4",
     "@igo2/context": "1.15.4",
-    "rxjs": "^7.5.6"
-  },
-  "optionalDependencies": {
     "jspdf": "^2.5.1",
-    "jspdf-autotable": "^3.5.29"
+    "jspdf-autotable": "^3.5.29",
+    "rxjs": "^7.8.0"
   },
   "dependencies": {
-    "tslib": "^2.4.0"
+    "tslib": "^2.6.0"
   },
   "engines": {
-    "node": ">=16.19.0"
+    "node": ">=18.10.0"
   }
 }
