@@ -5,8 +5,7 @@ import {
   Self,
   OnInit,
   HostListener
-} from '@angular/core'
-  ;
+} from '@angular/core';
 import olLayerVectorTile from 'ol/layer/VectorTile';
 import olLayerVector from 'ol/layer/Vector';
 
@@ -26,7 +25,7 @@ import * as OlGeom from 'ol/geom';
 import { EntityStore } from '@igo2/common';
 import { FeatureDataSource } from '../../datasource/shared/datasources/feature-datasource';
 import { VectorLayer, Layer, VectorTileLayer } from '../../layer/shared/layers';
-import { take } from 'rxjs/operators';
+import { first } from 'rxjs/operators';
 import { tryBindStoreLayer } from '../../feature/shared/feature.utils';
 import { FeatureStore } from '../../feature/shared/store';
 import { FeatureMotion } from '../../feature/shared/feature.enums';
@@ -36,6 +35,7 @@ import { unByKey } from 'ol/Observable';
 import RenderFeature from 'ol/render/Feature';
 import { StyleByAttribute } from '../../style/shared/vector/vector-style.interface';
 import { hoverFeatureMarkerStyle } from '../../style/shared/feature/feature-style';
+import { SubjectStatus } from '@igo2/utils';
 
 /**
  * This directive makes the mouse coordinate trigger a reverse search on available search sources.
@@ -109,10 +109,12 @@ export class HoverFeatureDirective implements OnInit, OnDestroy {
     this.subscribeToPointerStore();
     this.listenToMapClick();
 
-    this.map.status$.pipe(take(1)).subscribe(() => {
-      this.store = new FeatureStore<Feature>([], { map: this.map });
-      this.initStore();
-    });
+    this.map.status$
+      .pipe(first(status => status === SubjectStatus.Done))
+      .subscribe(() => {
+        this.store = new FeatureStore<Feature>([], { map: this.map });
+        this.initStore();
+      });
 
     // To handle context change without using the contextService.
     this.layers$$ = this.map.layers$.subscribe((layers: Layer[]) => {
