@@ -1,14 +1,14 @@
 import { NgModule } from '@angular/core';
-import { MatCardModule } from '@angular/material/card';
 
 import { IgoAuthModule } from '@igo2/auth';
 
 import { AppAuthFormComponent } from './auth-form.component';
 import { AppAuthFormRoutingModule } from './auth-form-routing.module';
+import { SharedModule } from '../../shared/shared.module';
 
 @NgModule({
   declarations: [AppAuthFormComponent],
-  imports: [AppAuthFormRoutingModule, MatCardModule, IgoAuthModule.forRoot()],
+  imports: [AppAuthFormRoutingModule, SharedModule, IgoAuthModule.forRoot()],
   exports: [AppAuthFormComponent]
 })
 export class AppAuthFormModule {}
