@@ -1,18 +1,16 @@
 import { NgModule } from '@angular/core';
-import { MatButtonModule } from '@angular/material/button';
-import { MatCardModule } from '@angular/material/card';
 
 import { IgoActionModule, IgoContextMenuModule } from '@igo2/common';
 
 import { AppActionComponent } from './action.component';
 import { AppActionRoutingModule } from './action-routing.module';
+import { SharedModule } from '../../shared/shared.module';
 
 @NgModule({
   declarations: [AppActionComponent],
   imports: [
+    SharedModule,
     AppActionRoutingModule,
-    MatButtonModule,
-    MatCardModule,
     IgoActionModule,
     IgoContextMenuModule
   ],
