@@ -1,6 +1,4 @@
 import { NgModule } from '@angular/core';
-import { MatButtonModule } from '@angular/material/button';
-import { MatCardModule } from '@angular/material/card';
 
 import { IgoDynamicComponentModule } from '@igo2/common';
 
@@ -10,6 +8,7 @@ import {
   AppDynamicComponentComponent
 } from './dynamic-component.component';
 import { AppDynamicComponentRoutingModule } from './dynamic-component-routing.module';
+import { SharedModule } from '../../shared/shared.module';
 
 @NgModule({
     declarations: [
@@ -18,9 +17,8 @@ import { AppDynamicComponentRoutingModule } from './dynamic-component-routing.mo
         AppExplanationComponent
     ],
     imports: [
+        SharedModule,
         AppDynamicComponentRoutingModule,
-        MatButtonModule,
-        MatCardModule,
         IgoDynamicComponentModule
     ],
     exports: [AppDynamicComponentComponent]
