@@ -1,16 +1,16 @@
 import { NgModule } from '@angular/core';
-import { MatCardModule } from '@angular/material/card';
 
 import { IgoTableModule } from '@igo2/common';
 
 import { AppTableComponent } from './table.component';
 import { AppTableRoutingModule } from './table-routing.module';
+import { SharedModule } from '../../shared/shared.module';
 
 @NgModule({
   declarations: [AppTableComponent],
   imports: [
+    SharedModule,
     AppTableRoutingModule,
-    MatCardModule,
     IgoTableModule
   ],
   exports: [AppTableComponent]
