@@ -1,12 +1,12 @@
 import { NgModule } from '@angular/core';
-import { MatCardModule } from '@angular/material/card';
 
 import { AppMediaComponent } from './media.component';
 import { AppMediaRoutingModule } from './media-routing.module';
+import { SharedModule } from '../../shared/shared.module';
 
 @NgModule({
   declarations: [AppMediaComponent],
-  imports: [AppMediaRoutingModule, MatCardModule],
+  imports: [AppMediaRoutingModule, SharedModule],
   exports: [AppMediaComponent]
 })
 export class AppMediaModule {}
