@@ -1,18 +1,16 @@
 import { NgModule } from '@angular/core';
-import { MatButtonModule } from '@angular/material/button';
-import { MatCardModule } from '@angular/material/card';
 
 import { IgoMessageModule } from '@igo2/core';
 
 import { AppMessageComponent } from './message.component';
 import { AppMessageRoutingModule } from './message-routing.module';
+import { SharedModule } from '../../shared/shared.module';
 
 @NgModule({
   declarations: [AppMessageComponent],
   imports: [
     AppMessageRoutingModule,
-    MatCardModule,
-    MatButtonModule,
+    SharedModule,
     IgoMessageModule.forRoot()
   ],
   exports: [AppMessageComponent]
