@@ -1,16 +1,16 @@
 import { NgModule } from '@angular/core';
-import { MatCardModule } from '@angular/material/card';
 
 import { IgoMapModule, IgoDrawModule } from '@igo2/geo';
 
 import { AppDrawComponent } from './draw.component';
 import { AppDrawRoutingModule } from './draw-routing.module';
+import { SharedModule } from '../../shared/shared.module';
 
 @NgModule({
   declarations: [AppDrawComponent],
   imports: [
     AppDrawRoutingModule,
-    MatCardModule,
+    SharedModule,
     IgoMapModule,
     IgoDrawModule
   ],
