@@ -1,23 +1,17 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { MatButtonModule } from '@angular/material/button';
-import { MatCardModule } from '@angular/material/card';
-import { MatIconModule } from '@angular/material/icon';
 
 import { IgoPanelModule, IgoEntityTableModule } from '@igo2/common';
 import { IgoMapModule, IgoFeatureModule } from '@igo2/geo';
 
 import { AppFeatureComponent } from './feature.component';
 import { AppFeatureRoutingModule } from './feature-routing.module';
+import { SharedModule } from '../../shared/shared.module';
 
 @NgModule({
   declarations: [AppFeatureComponent],
   imports: [
+    SharedModule,
     AppFeatureRoutingModule,
-    CommonModule,
-    MatCardModule,
-    MatButtonModule,
-    MatIconModule,
     IgoPanelModule,
     IgoEntityTableModule,
     IgoMapModule,
