@@ -1,6 +1,4 @@
 import { NgModule } from '@angular/core';
-import { MatButtonModule } from '@angular/material/button';
-import { MatCardModule } from '@angular/material/card';
 
 import { IgoMessageModule } from '@igo2/core';
 import {
@@ -12,13 +10,13 @@ import {
 
 import { AppImportExportComponent } from './import-export.component';
 import { AppImportExportRoutingModule } from './import-export-routing.module';
+import { SharedModule } from '../../shared/shared.module';
 
 @NgModule({
   declarations: [AppImportExportComponent],
   imports: [
     AppImportExportRoutingModule,
-    MatCardModule,
-    MatButtonModule,
+    SharedModule,
     IgoMessageModule,
     IgoMapModule,
     IgoStyleModule,
