@@ -1,16 +1,16 @@
 import { NgModule } from '@angular/core';
-import { MatCardModule } from '@angular/material/card';
 
 import { IgoMapModule, IgoMeasureModule } from '@igo2/geo';
 
 import { AppMeasureComponent } from './measure.component';
 import { AppMeasureRoutingModule } from './measure-routing.module';
+import { SharedModule } from '../../shared/shared.module';
 
 @NgModule({
   declarations: [AppMeasureComponent],
   imports: [
     AppMeasureRoutingModule,
-    MatCardModule,
+    SharedModule,
     IgoMapModule,
     IgoMeasureModule
   ],
