@@ -1,21 +1,17 @@
 import { NgModule } from '@angular/core';
-import { MatButtonModule } from '@angular/material/button';
-import { MatCardModule } from '@angular/material/card';
-import { MatIconModule } from '@angular/material/icon';
 
 import { IgoPanelModule } from '@igo2/common';
 import { IgoMapModule, IgoFilterModule } from '@igo2/geo';
 
 import { AppOgcFilterComponent } from './ogc-filter.component';
 import { AppOgcFilterRoutingModule } from './ogc-filter-routing.module';
+import { SharedModule } from '../../shared/shared.module';
 
 @NgModule({
   declarations: [AppOgcFilterComponent],
   imports: [
     AppOgcFilterRoutingModule,
-    MatCardModule,
-    MatButtonModule,
-    MatIconModule,
+    SharedModule,
     IgoPanelModule,
     IgoMapModule,
     IgoFilterModule
