@@ -1,8 +1,4 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { MatButtonModule } from '@angular/material/button';
-import { MatCardModule } from '@angular/material/card';
-import { MatIconModule } from '@angular/material/icon';
 
 import { IgoPanelModule } from '@igo2/common';
 import {
@@ -14,15 +10,13 @@ import {
 
 import { AppQueryComponent } from './query.component';
 import { AppQueryRoutingModule } from './query-routing.module';
+import { SharedModule } from '../../shared/shared.module';
 
 @NgModule({
   declarations: [AppQueryComponent],
   imports: [
     AppQueryRoutingModule,
-    CommonModule,
-    MatCardModule,
-    MatButtonModule,
-    MatIconModule,
+    SharedModule,
     IgoPanelModule,
     IgoMapModule,
     IgoOverlayModule,
