@@ -1,10 +1,4 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-import { MatButtonModule } from '@angular/material/button';
-import { MatCardModule } from '@angular/material/card';
-import { MatIconModule } from '@angular/material/icon';
-import { MatTooltipModule } from '@angular/material/tooltip';
 
 import { IgoMessageModule } from '@igo2/core';
 
@@ -33,16 +27,13 @@ import { IgoAppSearchModule } from '@igo2/integration';
 
 import { AppSearchComponent } from './search.component';
 import { AppSearchRoutingModule } from './search-routing.module';
+import { SharedModule } from '../../shared/shared.module';
 
 @NgModule({
   declarations: [AppSearchComponent],
   imports: [
-    CommonModule,
+    SharedModule,
     AppSearchRoutingModule,
-    MatCardModule,
-    MatButtonModule,
-    MatIconModule,
-    MatTooltipModule,
     IgoMessageModule.forRoot(),
     IgoPanelModule,
     IgoMapModule,
