@@ -6,12 +6,12 @@ import { IgoMapModule } from '@igo2/geo';
 
 import { AppSimpleMapComponent } from './simple-map.component';
 import { AppSimpleMapRoutingModule } from './simple-map-routing.module';
-import { CommonModule } from '@angular/common';
+import { SharedModule } from '../../shared/shared.module';
 
 @NgModule({
   declarations: [AppSimpleMapComponent],
   imports: [
-    CommonModule,
+    SharedModule,
     AppSimpleMapRoutingModule,
     MatCardModule,
     MatButtonModule,
