@@ -1,8 +1,4 @@
-import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { MatButtonModule } from '@angular/material/button';
-import { MatCardModule } from '@angular/material/card';
-import { MatIconModule } from '@angular/material/icon';
 
 import { IgoPanelModule, IgoFormModule } from '@igo2/common';
 import { IgoMapModule, IgoFilterModule, IgoQueryModule, IgoFeatureModule, IgoFeatureDetailsModule } from '@igo2/geo';
@@ -10,14 +6,13 @@ import { IgoMessageModule } from '@igo2/core';
 
 import { AppSpatialFilterComponent } from './spatial-filter.component';
 import { AppSpatialFilterRoutingModule } from './spatial-filter-routing.module';
+import { SharedModule } from '../../shared/shared.module';
 
 @NgModule({
   declarations: [AppSpatialFilterComponent],
   imports: [
     AppSpatialFilterRoutingModule,
-    MatCardModule,
-    MatButtonModule,
-    MatIconModule,
+    SharedModule,
     IgoPanelModule,
     IgoMapModule,
     IgoMessageModule,
@@ -25,8 +20,7 @@ import { AppSpatialFilterRoutingModule } from './spatial-filter-routing.module';
     IgoFeatureModule,
     IgoFeatureDetailsModule,
     IgoFilterModule,
-    IgoFormModule,
-    CommonModule
+    IgoFormModule
   ],
   exports: [AppSpatialFilterComponent]
 })
