@@ -13,12 +13,24 @@ module.exports = function (config) {
       require('@angular-devkit/build-angular/plugins/karma')
     ],
     client: {
+      jasmine: {
+        // you can add configuration options for Jasmine here
+        // the possible options are listed at https://jasmine.github.io/api/edge/Configuration.html
+        // for example, you can disable the random execution with `random: false`
+        // or set a specific seed with `seed: 4321`
+      },
       clearContext: false // leave Jasmine Spec Runner output visible in browser
     },
+    jasmineHtmlReporter: {
+      suppressAll: true // removes the duplicated traces
+    },
     coverageReporter: {
-      dir: require('path').join(__dirname, '../../coverage'),
-      reporters: ['html', 'lcovonly'],
-      fixWebpackSourcePaths: true
+      dir: require('path').join(__dirname, '../../coverage/auth'),
+      subdir: '.',
+      reporters: [
+        { type: 'html', subdir: 'report-html' },
+        { type: 'text-summary', file: 'report-summary.txt' },
+      ],
     },
     reporters: ['progress', 'kjhtml'],
     port: 9876,
