@@ -83,7 +83,7 @@ export class AuthFormComponent implements OnInit {
 
   @Output() login: EventEmitter<boolean> = new EventEmitter<boolean>();
 
-  public options: AuthOptions;
+  public options?: AuthOptions;
   public user;
 
   public visible = true;
@@ -96,7 +96,7 @@ export class AuthFormComponent implements OnInit {
     private config: ConfigService,
     @Optional() private router: Router
   ) {
-    this.options = this.config.getConfig('auth') || {};
+    this.options = this.config.getConfig('auth');
     this.visible = Object.getOwnPropertyNames(this.options).length !== 0;
   }
 
@@ -115,18 +115,18 @@ export class AuthFormComponent implements OnInit {
     this.auth.logout().subscribe(() => {
       this.user = undefined;
       if (this.router) {
-        if (this.options.logoutRoute) {
-          this.router.navigate([this.options.logoutRoute]);
-        } else if (this.options.homeRoute) {
-          this.router.navigate([this.options.homeRoute]);
+        if (this.options?.logoutRoute) {
+          this.router.navigate([this.options?.logoutRoute]);
+        } else if (this.options?.homeRoute) {
+          this.router.navigate([this.options?.homeRoute]);
         }
       }
     });
   }
 
   public home() {
-    if (this.router && this.options.homeRoute) {
-      this.router.navigate([this.options.homeRoute]);
+    if (this.router && this.options?.homeRoute) {
+      this.router.navigate([this.options?.homeRoute]);
     }
   }
 
@@ -149,8 +149,8 @@ export class AuthFormComponent implements OnInit {
       .subscribe((changeEvent: any) => {
         if (changeEvent.url) {
           const currentRoute = changeEvent.url;
-          const logoutRoute = this.options.logoutRoute;
-          const loginRoute = this.options.loginRoute;
+          const logoutRoute = this.options?.logoutRoute;
+          const loginRoute = this.options?.loginRoute;
 
           this.isLogoutRoute = currentRoute === logoutRoute;
           this.isLoginRoute = currentRoute === loginRoute;
