@@ -26,19 +26,19 @@ import { UserButtonComponent } from './user-button/user-button.component';
 @NgModule({
   imports: [
     CommonModule,
-    IgoLanguageModule,
-    IgoConfirmDialogModule,
-    IgoStopPropagationModule,
-    IgoAuthModule,
     FormsModule,
-    MatIconModule,
+    IgoAuthModule,
+    IgoConfirmDialogModule,
+    IgoLanguageModule,
+    IgoStopPropagationModule,
     MatButtonModule,
-    MatSelectModule,
-    MatOptionModule,
-    MatTooltipModule,
-    MatFormFieldModule,
     MatDialogModule,
-    MatInputModule
+    MatFormFieldModule,
+    MatIconModule,
+    MatInputModule,
+    MatOptionModule,
+    MatSelectModule,
+    MatTooltipModule,
   ],
   exports: [BookmarkButtonComponent, PoiButtonComponent, UserButtonComponent, BookmarkDialogComponent],
   declarations: [
