@@ -5,5 +5,5 @@ export interface Version {
 
 export const version: Version = {
   lib: '1.15.4',
-  releaseDate: 1688557109892
+  releaseDate: 1689596479901
 };
