@@ -9,7 +9,8 @@ import { GlobalConfig, ToastrModule } from 'ngx-toastr';
       positionClass: 'toast-bottom-right',
       timeOut: 10000,
       extendedTimeOut: 10000,
-      messageClass: 'toast-message mat-typography',
+      titleClass: 'mat-subtitle-2',
+      messageClass: 'toast-message',
       closeButton: true,
       progressBar: true,
       enableHtml: true,
