@@ -34,24 +34,24 @@ import { SaveFeatureDialogComponent } from './save-feature-dialog.component';
 @NgModule({
   imports: [
     CommonModule,
-    MatBadgeModule,
-    MatTooltipModule,
-    MatIconModule,
-    MatListModule,
-    MatButtonModule,
-    MatTabsModule,
-    MatDialogModule,
     IgoCollapsibleModule,
-    IgoListModule,
-    IgoStopPropagationModule,
     IgoLanguageModule,
+    IgoListModule,
     IgoMatBadgeIconModule,
     IgoMetadataModule,
-    MatFormFieldModule,
-    ReactiveFormsModule,
-    MatInputModule,
-    MatSelectModule,
+    IgoStopPropagationModule,
     MatAutocompleteModule,
+    MatBadgeModule,
+    MatButtonModule,
+    MatDialogModule,
+    MatFormFieldModule,
+    MatIconModule,
+    MatInputModule,
+    MatListModule,
+    MatSelectModule,
+    MatTabsModule,
+    MatTooltipModule,
+    ReactiveFormsModule,
   ],
   exports: [
     SearchResultsComponent,
